beginning of decoration for grass tiles
also things for gabe

+++ TwitterIsland/Assets/Scripts/SerializationHelper.cs
@@ -7,14 +7,14 @@ public class SerializationHelper {
 
     public static string VectorToJson(Vector3 vec)
     {
-        float[] arr = { vec.x, vec.y, vec.z };
+        float[] arr = { vec.x, (int)vec.y, vec.z };
         return JsonConvert.SerializeObject(arr);
     }
 
     public static Vector3 JsonToVector(string json)
     {
         float[] arr = JsonConvert.DeserializeObject<float[]>(json);
-        return new Vector3(arr[0], arr[1], arr[2]);
+        return new Vector3(arr[0], (int)arr[1], arr[2]);
     }
 
 }
